+++ src/main/java/roomescape/dao/member/JdbcMemberDao.java
@@ -1,4 +1,4 @@
-package roomescape.dao.meber;
+package roomescape.dao.member;
 
 import java.util.List;
 import java.util.Optional;
