+++ src/test/java/roomescape/dao/member/JdbcMemberDaoTest.java
@@ -1,4 +1,4 @@
-package roomescape.dao.meber;
+package roomescape.dao.member;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
